requirements.txt
openai
phython-dotenv 